% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterTFs.R
\name{ClusterTFs}
\alias{ClusterTFs}
\title{Cluster transcription factors}
\usage{
ClusterTFs(Merged_Peak_Frame)
}
\arguments{
\item{Merged_Peak_Frame}{A dataframe containing merged TF peak data with at least 4 columns. This is the standard output from the SMART_BED_MERGE function in GenomeTools2020 repo. Columns 1-3 should contain the chromosome, start, and stop position of each merged peak region. Column 4 should contain a string that represents comma separated list of all TFs present at each position. Each row should contain data for each merged peak region.}
}
\value{
Returns a distance matrix indicating the clustering distance for each TF pair. The number of columns and rows is equal to the number of unique TFs included in the Merged_Peak_Frame. This matrix can be provided to any heatmap function desired.
}
\description{
Cluster transcription factors
}
\examples{
data(DLPFC_Example_062120)
ClusterTFs(test)
}

